+++ src/app/layout.tsx
@@ -1,0 +1,8 @@
+// layout.tsx
+interface Props {
+  children: React.ReactNode;
+}
+
+export default function RootLayout({ children }: Props) {
+  return <>{children}</>;
+}
